seperated soup functions into Content folder, start of dungeon scrape
